requirements.txt
# ======================================
# PRODUCTION DEPENDENCIES
# Used in: dev, staging, prod environments
# ======================================

# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.7.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9

# HTTP requests
httpx==0.25.2
requests==2.31.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
bcrypt==4.1.2

# Monitoring
prometheus-client==0.19.0